+++ ubo_app/services/0-ip/reducer.py
@@ -1,26 +1,32 @@
 # ruff: noqa: D100, D101, D102, D103, D104, D107, N999
 from __future__ import annotations
 
+from dataclasses import replace
+
 from redux import (
     Action,
     BaseEvent,
-    Immutable,
+    CompleteReducerResult,
     InitAction,
     InitializationActionError,
     ReducerResult,
 )
 
-
-class IPState(Immutable):
-    ...
+from ubo_app.store.ip import IpState, IpUpdateRequestAction, IpUpdateRequestEvent
 
 
 def reducer(
-    state: IPState | None,
+    state: IpState | None,
     action: Action,
-) -> ReducerResult[IPState, Action, BaseEvent]:
+) -> ReducerResult[IpState, Action, BaseEvent]:
     if state is None:
         if isinstance(action, InitAction):
-            return IPState()
+            return IpState(interfaces=[])
         raise InitializationActionError
+
+    if isinstance(action, IpUpdateRequestAction):
+        return CompleteReducerResult(
+            state=replace(state, status=None) if action.reset else state,
+            events=[IpUpdateRequestEvent()],
+        )
     return state